+++ bin/api-newsletter-add.js
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-import apiNewsletterAdd from '../src/hooks/init.js'
-
-await apiNewsletterAdd()