+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+## [3.3.2] - 2024-05-07
+
+### Changed
+
+- Oclif Init Hook: Request User email text updated
+- Oclif Init Hook: Now only triggers on our plugin commands.
+
+### Fixed
+
+- Generate: The component name did not replace the 'component-name' term in the JS file name.
+- Dev: Error when specifying a component name
+
+### Removed
+
+- NPM Install Hook: Request User email for DevKit Newsletter Subscription #417
+
+
 ## [3.3.1] - 2024-05-07
 
 ### Fixed
@@ -12,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Reference to favicon component
+
 
 ## [3.3.0] - 2024-05-06